requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
decorator==4.4.2
Django==3.0
django-appconf==1.0.4
django-classy-tags==2.0.0
django-compressor==2.4
django-js-asset==1.2.2
django-mptt==0.11.0
django-sass-processor==0.8
djangorestframework==3.11.1
FeinCMS==1.18.0
future==0.18.2
geocoder==1.38.1
geographiclib==1.50
geopy==2.0.0
idna==2.10
Pillow==7.2.0
pytz==2020.1
ratelim==0.1.6
rcssmin==1.0.6
requests==2.24.0
rjsmin==1.1.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.10
